[folder org] clearer distinction btwn history builtin/substitution
