+++ srcs/history/builtin/history_del.c
