+++ srcs/history/substitution/get_entry_word.c
