+++ srcs/history/substitution/substitute_str.c
