+++ srcs/history/substitution/word_designators.c
