+++ srcs/read_input/editor/editor.c
@@ -11,7 +11,7 @@
 # include "read_input/event_callbacks/event_callback_def.h"
 # include "read_input/termcap/init_deinit.h"
 
-# include "history/history.h"
+# include "history/history_def.h"
 # include "variable.h"
 
 t_editor *get_editor()
