+++ srcs/read_input/event_callbacks/history.c
@@ -1,5 +1,5 @@
 #include "event_callback_def.h"
-#include "history/history.h"
+#include "history/history_def.h"
 #include "abstract_list.h"
 #include <libft.h>
